+++ packages/wow-ui/src/components/Picker/SingleDatePicker.tsx
@@ -6,10 +6,10 @@ import type { PropsBase, PropsSingle } from "react-day-picker";
 import { DayPicker } from "react-day-picker";
 
 import DateDropDown from "@/components/Picker/DateDropDown";
-import { pickerButtonStyle } from "@/components/Picker/pickerButtonStyle.css";
-import pickerClassNames from "@/components/Picker/pickerClassNames";
-import pickerComponents from "@/components/Picker/pickerComponents";
 import { usePicker } from "@/components/Picker/PickerContext";
+import { pickerButtonStyle } from "@/components/Picker/pickerProps/pickerButtonStyle.css";
+import pickerClassNames from "@/components/Picker/pickerProps/pickerClassNames";
+import pickerComponents from "@/components/Picker/pickerProps/pickerComponents";
 import { changeDateToString } from "@/utils/changeDateToString";
 
 export type DatePickerProps = Omit<PropsBase, "mode"> &
