+++ packages/wow-ui/src/components/Picker/TimePicker.tsx
@@ -3,12 +3,12 @@ import { forwardRef, useMemo, useState } from "react";
 
 import DropDown from "@/components/DropDown";
 import DropDownOption from "@/components/DropDown/DropDownOption";
-import { pickerButtonStyle } from "@/components/Picker/pickerButtonStyle.css";
 import type {
   PickerContextProps,
   Time,
 } from "@/components/Picker/PickerContext";
 import { usePicker } from "@/components/Picker/PickerContext";
+import { pickerButtonStyle } from "@/components/Picker/pickerProps/pickerButtonStyle.css";
 
 interface TimePickerProps
   extends Partial<
